chore: updated pages response
