+++ packages/types/src/pages.d.ts
@@ -11,10 +11,10 @@ export type TPage = {
   id: string | undefined;
   is_favorite: boolean;
   is_locked: boolean;
-  labels: string[] | undefined;
+  label_ids: string[] | undefined;
   name: string | undefined;
   owned_by: string | undefined;
-  project: string | undefined;
+  project_ids: string[] | undefined;
   updated_at: Date | undefined;
   updated_by: string | undefined;
   workspace: string | undefined;
